Make it work with signal-cli 0.10.3; custom headers; names from DBUS
Most important change: compatibility with v0.10.3:
As of signal-cli v0.10.3 the V2 API does not or no longer an additional
argument "mentionList"; but instead a dict "extras" with attachments,
mention-list and other things.

Also: the format of the attachment array has changed, it is now a dict
(much easier than before if I read the signalmail code correctly)

Same holds for the mention-list, but this is fixed in a later commit.

Further changes:
- As I plan to pipe this through (sieve-)filters it is handy to add custom
headers which are customizable in the config-file.
- Some interpolation of sender-/group- id/name is added to subject and
  from-field, as well as to the headers
- parsing of contacts has been moved out of the event-handler. The
  configured names are now injected into signal-cli with
  setContactName()
- the group-name is determined if a group is detected
- the signalgroupid has been removed from the config-file, it wasn't
  used anywhere.