Cleanup replacements, make time-format and heading configurable, add signature.
The signature is in so far important as on a reply the old signature is
remove by most email clients. If we now forward a Signal-group to a
mailing list, the mailing list will add its own fancy heading below the
original signature. However, the mailing list signature-separator is
normally not detected by the clients and hence we get a cascade of
mailing-list signature in a message thread. This should be avoided.